requirements.txt
asgiref==3.4.1
cfgv==3.3.1
distlib==0.3.4
Django==4.0.1
django-materializecss-form==1.1.17
django-truncate==0.1
djangorestframework==3.13.1
filelock==3.4.2
identify==2.4.8
mysqlclient==2.1.0
nodeenv==1.6.0
platformdirs==2.4.1
pre-commit==2.17.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.13.0
django-import-export==2.7.1
